+++ client/tsconfig.json
@@ -8,5 +8,10 @@
 		"sourceMap": true
 	},
 	"include": ["src"],
-	"exclude": ["node_modules", ".vscode-test"]
+	"exclude": ["node_modules", ".vscode-test"],
+	"references": [
+		{
+			"path": "../server"
+		}
+	]
 }
