+++ server/src/dbc/node.ts
@@ -1,0 +1,30 @@
+/**
+ * Copyright (C) 2021 Landon Harris
+ * This program is free software; you can redistribute it and/or 
+ * modify it under the terms of the GNU General Public License as 
+ * published by the Free Software Foundation; version 2.
+ * 
+ * This program is distributed in the hope that it will be useful, 
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of 
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License 
+ * along with this program. If not, see 
+ * <https://www.gnu.org/licenses/old-licenses/gpl-2.0-standalone.html>.
+*/
+
+import { Attribute } from "./attributes";
+
+export class Node{
+    public constructor(name: string){
+        this.name = name;
+        this.comment = "";
+        this.attributes = new Map();
+    }
+
+    public name: string;
+    public comment: string;
+    public attributes: Map<string,Attribute>;
+}
+
